Update Feature Extractor module
This module was modified to extract features in a generic way, regardless of the argument mining task to be performed later.
