+++ code/FeatureExtractor/Resources/config/params.json
@@ -1,5 +1,4 @@
 {
-    "extractionMode": "detection",
     "language": "es",
     "linkers": {
         "en": {
