+++ code/FeatureExtractor/src/es/uam/irg/nlp/am/Program.java
@@ -36,16 +36,15 @@ public class Program {
         // Program hyperparameters from JSON config file
         Map<String, Object> params = InitParams.readInitParams();
         String language = (String) params.get("language");
-        String extractionMode = (String) params.get("extractionMode");
         Map<String, HashSet<String>> linkers = (Map<String, HashSet<String>>) params.get("linkers");
         HashSet<String> validLinkers = linkers.get("validLinkers");
         HashSet<String> invalidLinkers = linkers.get("invalidLinkers");
-        System.out.format(">> Analysis language: %s, Extraction mode: %s, Valid linkers: %s, Invalid linkers: %s\n",
-                language, extractionMode, validLinkers, invalidLinkers);
+        System.out.format(">> Analysis language: %s, Valid linkers: %s, Invalid linkers: %s\n",
+                language, validLinkers, invalidLinkers);
 
         // Run program
         FeatureExtractor miner = new FeatureExtractor(language, validLinkers, invalidLinkers);
-        boolean result = miner.runProgram(extractionMode);
+        boolean result = miner.runProgram();
 
         if (result) {
             System.out.println(">> The Feature Extractor engine was executed correctly.");
